backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements.txt requirements.in requirements.private.in
#
arrow==0.15.4
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask
flask-cors==3.0.8
flask==1.1.1
geographiclib==1.50       # via geopy
geopy==1.20.0
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.10.3            # via flask
markupsafe==1.1.1         # via jinja2
numpy==1.17.4
psycopg2-binary==2.8.4
python-dateutil==2.8.1    # via arrow
requests==2.22.0
git+git://github.com/camenpihor/rogue-sky.git#egg=rogue-sky
six==1.13.0               # via flask-cors, python-dateutil
sqlparse==0.3.0
urllib3==1.25.7           # via requests
werkzeug==0.16.0          # via flask
